+++ app/routes/home/partials/what-we-offer.partial.tsx
@@ -1,16 +1,10 @@
-import { WhatWeOfferMobile } from '../components/we-offer-tabs/mobile.component';
-import { WhatWeOfferDesktop } from '../components/we-offer-tabs/desktop.component';
+import { WhatWeOfferTabs } from '../components/we-offer-tabs/what-we-offer-tabs.component';
 
 export function WhatWeOfferSection() {
   return (
     <section className='w-full py-24 md:pt-38 md:pb-23 bg-navy relative z-30'>
       <div className='max-w-screen-content mx-auto flex flex-col items-center gap-12'>
-        <div className='md:hidden'>
-          <WhatWeOfferMobile />
-        </div>
-        <div className='hidden md:block'>
-          <WhatWeOfferDesktop />
-        </div>
+        <WhatWeOfferTabs />
       </div>
     </section>
   );
